+++ eda-modbus-bridge.mjs
@@ -3,7 +3,7 @@ import expressWinston from 'express-winston'
 import mqtt from 'mqtt'
 import yargs from 'yargs'
 import ModbusRTU from 'modbus-serial'
-import { getFlagStatus, root, setFlagStatus, setSetting, summary } from './app/http.mjs'
+import { getMode, root, setMode, setSetting, summary } from './app/http.mjs'
 import {
     publishValues,
     subscribeToChanges,
@@ -125,11 +125,11 @@ const argv = yargs(process.argv.slice(2))
         httpServer.get('/summary', (req, res) => {
             return summary(modbusClient, req, res)
         })
-        httpServer.get('/mode/:flag', (req, res) => {
-            return getFlagStatus(modbusClient, req, res)
+        httpServer.get('/mode/:mode', (req, res) => {
+            return getMode(modbusClient, req, res)
         })
-        httpServer.post('/mode/:flag', (req, res) => {
-            return setFlagStatus(modbusClient, req, res)
+        httpServer.post('/mode/:mode', (req, res) => {
+            return setMode(modbusClient, req, res)
         })
         httpServer.post('/setting/:setting/:value', (req, res) => {
             return setSetting(modbusClient, req, res)